pyproject.toml
[project]
name = "data-integration-benchmark"
version = "1.0.0"
description = "Green Agent for TPC-DI Data Integration Benchmark"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "uvicorn>=0.38.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "fastmcp>=2.0.0",
    "fastapi>=0.115.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
